Split tests per flavor
