+++ purchases/src/testDefaults/kotlin/com/revenuecat/purchases/PurchasesCoroutinesTest.kt
@@ -81,69 +81,4 @@ internal class PurchasesCoroutinesTest : BasePurchasesTest() {
     }
 
     // endregion
-
-    // region awaitOfferings
-    @Test
-    fun `retrieve offerings - Success`() = runTest {
-        mockOfferingsManagerGetOfferings()
-
-        val result = purchases.awaitOfferings()
-
-        verify(exactly = 1) {
-            mockOfferingsManager.getOfferings(
-                appUserId,
-                any(),
-                any(),
-                any(),
-            )
-        }
-        assertThat(result).isNotNull
-    }
-
-    @Test
-    fun `retrieve offerings - Success - offerings match expectations`() = runTest {
-        val offerings = mockOfferingsManagerGetOfferings()
-
-        val result = purchases.awaitOfferings()
-
-        verify(exactly = 1) {
-            mockOfferingsManager.getOfferings(
-                appUserId,
-                any(),
-                any(),
-                any(),
-            )
-        }
-        assertThat(result).isNotNull
-        assertThat(result).isEqualTo(offerings)
-    }
-
-    @Test
-    fun `retrieve offerings - ConfigurationError`() = runTest {
-        val error = PurchasesError(PurchasesErrorCode.ConfigurationError, "Offerings config error")
-        mockOfferingsManagerGetOfferings(error)
-
-        var result: Offerings? = null
-        var exception: Throwable? = null
-        runCatching {
-            result = purchases.awaitOfferings()
-        }.onFailure {
-            exception = it
-        }
-
-        verify(exactly = 1) {
-            mockOfferingsManager.getOfferings(
-                appUserId,
-                any(),
-                any(),
-                any(),
-            )
-        }
-        assertThat(result).isNull()
-        assertThat(exception).isNotNull
-        assertThat(exception).isInstanceOf(PurchasesException::class.java)
-        assertThat((exception as PurchasesException).code).isEqualTo(PurchasesErrorCode.ConfigurationError)
-    }
-
-    // endregion
 }
